+++ NEWS.md
@@ -1,3 +1,7 @@
+# bcgwcat 0.6.1
+* Allow water types to use Meas Alk if HCO3 is missing (identifyed by HCO3* in the returned water type).
+* Water type also returns a check column `mission_ion` to identify if the water type is missing all anions or all cations.
+
 # bcgwcat 0.6.0
 * Rename to bcgwcat to reflect change in scope
   * Update docs etc. in light of this